+++ src/Features/Core/Portable/CodeLens/CodeLensReferencesService.cs
@@ -10,9 +10,11 @@ using System.Globalization;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.CodeAnalysis.Classification;
 using Microsoft.CodeAnalysis.FindSymbols;
 using Microsoft.CodeAnalysis.Host;
 using Microsoft.CodeAnalysis.LanguageService;
+using Microsoft.CodeAnalysis.Options;
 using Microsoft.CodeAnalysis.PooledObjects;
 using Microsoft.CodeAnalysis.Shared.Extensions;
 using Microsoft.CodeAnalysis.Text;
@@ -98,7 +100,7 @@ internal sealed class CodeLensReferencesService : ICodeLensReferencesService
             maxSearchResults, cancellationToken).ConfigureAwait(false);
     }
 
-    private static async Task<ReferenceLocationDescriptor> GetDescriptorOfEnclosingSymbolAsync(Solution solution, Location location, CancellationToken cancellationToken)
+    private static async Task<ReferenceLocationDescriptorAndDocument> GetDescriptorOfEnclosingSymbolAsync(Solution solution, Location location, CancellationToken cancellationToken)
     {
         var document = solution.GetDocument(location.SourceTree);
 
@@ -130,41 +132,36 @@ internal sealed class CodeLensReferencesService : ICodeLensReferencesService
         var spanStart = token.Span.Start - textLine.Span.Start;
         var line = textLine.ToString();
 
-        var beforeLine1 = textLine.LineNumber > 0 ? text.Lines[textLine.LineNumber - 1].ToString() : string.Empty;
-        var beforeLine2 = textLine.LineNumber - 1 > 0
-            ? text.Lines[textLine.LineNumber - 2].ToString()
-            : string.Empty;
-        var afterLine1 = textLine.LineNumber < text.Lines.Count - 1
-            ? text.Lines[textLine.LineNumber + 1].ToString()
-            : string.Empty;
-        var afterLine2 = textLine.LineNumber + 1 < text.Lines.Count - 1
-            ? text.Lines[textLine.LineNumber + 2].ToString()
-            : string.Empty;
+        var beforeLine1 = GetLineTextOrEmpty(text.Lines, textLine.LineNumber - 1);
+        var beforeLine2 = GetLineTextOrEmpty(text.Lines, textLine.LineNumber - 2);
+        var afterLine1 = GetLineTextOrEmpty(text.Lines, textLine.LineNumber + 1);
+        var afterLine2 = GetLineTextOrEmpty(text.Lines, textLine.LineNumber + 2);
         var referenceSpan = new TextSpan(spanStart, token.Span.Length);
 
         var symbol = semanticModel.GetDeclaredSymbol(node, cancellationToken);
         var glyph = symbol?.GetGlyph();
         var startLinePosition = location.GetLineSpan().StartLinePosition;
-        var documentId = solution.GetDocument(location.SourceTree)?.Id;
 
-        return new ReferenceLocationDescriptor(
-            longName,
-            semanticModel.Language,
-            glyph,
-            token.Span.Start,
-            token.Span.Length,
-            startLinePosition.Line,
-            startLinePosition.Character,
-            documentId.ProjectId.Id,
-            documentId.Id,
-            document.FilePath,
-            line.TrimEnd(),
-            referenceSpan.Start,
-            referenceSpan.Length,
-            beforeLine1.TrimEnd(),
-            beforeLine2.TrimEnd(),
-            afterLine1.TrimEnd(),
-            afterLine2.TrimEnd());
+        return new ReferenceLocationDescriptorAndDocument
+        {
+            Descriptor = new ReferenceLocationDescriptor(
+                longName,
+                semanticModel.Language,
+                glyph,
+                token.Span.Start,
+                token.Span.Length,
+                startLinePosition.Line,
+                startLinePosition.Character,
+                document.FilePath,
+                line.TrimEnd(),
+                referenceSpan.Start,
+                referenceSpan.Length,
+                beforeLine1.TrimEnd(),
+                beforeLine2.TrimEnd(),
+                afterLine1.TrimEnd(),
+                afterLine2.TrimEnd()),
+            DocumentId = document.Id
+        };
     }
 
     private static SyntaxNode GetEnclosingCodeElementNode(Document document, SyntaxToken token, ICodeLensDisplayInfoService langServices, CancellationToken cancellationToken)
@@ -199,7 +196,7 @@ internal sealed class CodeLensReferencesService : ICodeLensReferencesService
         return langServices.GetDisplayNode(node);
     }
 
-    public async Task<ImmutableArray<ReferenceLocationDescriptor>?> FindReferenceLocationsAsync(Solution solution, DocumentId documentId, SyntaxNode syntaxNode, CancellationToken cancellationToken)
+    public async Task<ImmutableArray<ReferenceLocationDescriptorAndDocument>?> FindReferenceLocationsAsync(Solution solution, DocumentId documentId, SyntaxNode syntaxNode, CancellationToken cancellationToken)
     {
         return await FindAsync(solution, documentId, syntaxNode,
             async progress =>
@@ -212,6 +209,137 @@ internal sealed class CodeLensReferencesService : ICodeLensReferencesService
 
                 return result.ToImmutableArray();
             }, onCapped: null, searchCap: 0, cancellationToken: cancellationToken).ConfigureAwait(false);
+    }
+
+    public async Task<ImmutableArray<ReferenceLocationDescriptor>> MapReferenceLocationsAsync(Solution solution, ImmutableArray<ReferenceLocationDescriptorAndDocument> referenceLocations, ClassificationOptions classificationOptions, CancellationToken cancellationToken)
+    {
+        using var _ = ArrayBuilder<ReferenceLocationDescriptor>.GetInstance(out var list);
+        foreach (var descriptorAndDocument in referenceLocations)
+        {
+            var document = await solution.GetDocumentAsync(descriptorAndDocument.DocumentId, includeSourceGenerated: true, cancellationToken).ConfigureAwait(false);
+            if (document == null)
+            {
+                continue;
+            }
+
+            var descriptor = descriptorAndDocument.Descriptor;
+            var span = new TextSpan(descriptor.SpanStart, descriptor.SpanLength);
+            var results = await SpanMappingHelper.TryGetMappedSpanResultAsync(document, [span], cancellationToken).ConfigureAwait(false);
+            if (results is null)
+            {
+                // for normal document, just add one as they are
+                list.Add(descriptor);
+                continue;
+            }
+
+            var mappedSpans = results.GetValueOrDefault();
+
+            // external component violated contracts. the mapper should preserve input order/count. 
+            // since we gave in 1 span, it should return 1 span back
+            Contract.ThrowIfTrue(mappedSpans.IsDefaultOrEmpty);
+
+            var result = mappedSpans[0];
+            if (result.IsDefault)
+            {
+                // it is allowed for mapper to return default 
+                // if it can't map the given span to any usable span
+                continue;
+            }
+
+            var excerpter = document.DocumentServiceProvider.GetService<IDocumentExcerptService>();
+            if (excerpter == null)
+            {
+                if (document.IsRazorSourceGeneratedDocument())
+                {
+                    // HACK: Razor doesn't have has a workspace level excerpt service, but if we just return a simple descriptor here,
+                    // the user at least sees something, can navigate, and Razor can improve this later if necessary. Until
+                    // https://github.com/dotnet/roslyn/issues/79699 is fixed this won't get hit anyway.
+                    list.Add(new ReferenceLocationDescriptor(
+                        descriptor.LongDescription,
+                        descriptor.Language,
+                        descriptor.Glyph,
+                        result.Span.Start,
+                        result.Span.Length,
+                        result.LinePositionSpan.Start.Line,
+                        result.LinePositionSpan.Start.Character,
+                        result.FilePath,
+                        descriptor.ReferenceLineText,
+                        descriptor.ReferenceStart,
+                        descriptor.ReferenceLength,
+                        "",
+                        "",
+                        "",
+                        ""));
+                }
+                continue;
+            }
+
+            var referenceExcerpt = await excerpter.TryExcerptAsync(document, span, ExcerptMode.SingleLine, classificationOptions, cancellationToken).ConfigureAwait(false);
+            var tooltipExcerpt = await excerpter.TryExcerptAsync(document, span, ExcerptMode.Tooltip, classificationOptions, cancellationToken).ConfigureAwait(false);
+
+            var (text, start, length) = GetReferenceInfo(referenceExcerpt, descriptor);
+            var (before1, before2, after1, after2) = GetReferenceTexts(referenceExcerpt, tooltipExcerpt, descriptor);
+
+            list.Add(new ReferenceLocationDescriptor(
+                descriptor.LongDescription,
+                descriptor.Language,
+                descriptor.Glyph,
+                result.Span.Start,
+                result.Span.Length,
+                result.LinePositionSpan.Start.Line,
+                result.LinePositionSpan.Start.Character,
+                result.FilePath,
+                text,
+                start,
+                length,
+                before1,
+                before2,
+                after1,
+                after2));
+        }
+
+        return list.ToImmutableAndClear();
+    }
+
+    private static (string text, int start, int length) GetReferenceInfo(ExcerptResult? reference, ReferenceLocationDescriptor descriptor)
+    {
+        if (reference.HasValue)
+        {
+            return (reference.Value.Content.ToString().TrimEnd(),
+                    reference.Value.MappedSpan.Start,
+                    reference.Value.MappedSpan.Length);
+        }
+
+        return (descriptor.ReferenceLineText, descriptor.ReferenceStart, descriptor.ReferenceLength);
+    }
+
+    private static (string before1, string before2, string after1, string after2) GetReferenceTexts(ExcerptResult? reference, ExcerptResult? tooltip, ReferenceLocationDescriptor descriptor)
+    {
+        if (reference == null || tooltip == null)
+        {
+            return (descriptor.BeforeReferenceText1, descriptor.BeforeReferenceText2, descriptor.AfterReferenceText1, descriptor.AfterReferenceText2);
+        }
+
+        var lines = tooltip.Value.Content.Lines;
+        var mappedLine = lines.GetLineFromPosition(tooltip.Value.MappedSpan.Start);
+        var index = mappedLine.LineNumber;
+        if (index < 0)
+        {
+            return (descriptor.BeforeReferenceText1, descriptor.BeforeReferenceText2, descriptor.AfterReferenceText1, descriptor.AfterReferenceText2);
+        }
+
+        return (GetLineTextOrEmpty(lines, index - 1), GetLineTextOrEmpty(lines, index - 2),
+                GetLineTextOrEmpty(lines, index + 1), GetLineTextOrEmpty(lines, index + 2));
+    }
+
+    private static string GetLineTextOrEmpty(TextLineCollection lines, int index)
+    {
+        if (index < 0 || index >= lines.Count)
+        {
+            return string.Empty;
+        }
+
+        return lines[index].ToString().TrimEnd();
     }
 
     private static ISymbol GetEnclosingMethod(SemanticModel semanticModel, Location location, CancellationToken cancellationToken)
