+++ src/Features/Core/Portable/CodeLens/ICodeLensReferencesService.cs
@@ -5,6 +5,7 @@
 using System.Collections.Immutable;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.CodeAnalysis.Classification;
 using Microsoft.CodeAnalysis.Host;
 
 namespace Microsoft.CodeAnalysis.CodeLens;
@@ -25,7 +26,12 @@ internal interface ICodeLensReferencesService : IWorkspaceService
     /// <summary>
     /// Given a document and syntax node, returns a collection of locations where the located node is referenced.
     /// </summary>
-    Task<ImmutableArray<ReferenceLocationDescriptor>?> FindReferenceLocationsAsync(Solution solution, DocumentId documentId, SyntaxNode? syntaxNode, CancellationToken cancellationToken);
+    Task<ImmutableArray<ReferenceLocationDescriptorAndDocument>?> FindReferenceLocationsAsync(Solution solution, DocumentId documentId, SyntaxNode? syntaxNode, CancellationToken cancellationToken);
+
+    /// <summary>
+    /// Maps the list of referenced locations through span mapping services to the final list of locations; locations may be in files that aren't regular documents.
+    /// </summary>
+    Task<ImmutableArray<ReferenceLocationDescriptor>> MapReferenceLocationsAsync(Solution solution, ImmutableArray<ReferenceLocationDescriptorAndDocument> referenceLocations, ClassificationOptions classificationOptions, CancellationToken cancellationToken);
 
     /// <summary>
     /// Given a document and syntax node, returns a collection of locations of methods that refer to the located node.
