+++ src/VisualStudio/Core/Test/Venus/DocumentService_IntegrationTests.vb
@@ -10,6 +10,7 @@ Imports System.Windows.Controls
 Imports System.Windows.Media
 Imports Microsoft.CodeAnalysis
 Imports Microsoft.CodeAnalysis.Classification
+Imports Microsoft.CodeAnalysis.CodeLens
 Imports Microsoft.CodeAnalysis.Collections
 Imports Microsoft.CodeAnalysis.CSharp.Syntax
 Imports Microsoft.CodeAnalysis.Diagnostics
@@ -136,7 +137,7 @@ class {|Definition:C1|}
 
             Using workspace = EditorTestWorkspace.Create(input, documentServiceProvider:=TestDocumentServiceProvider.Instance)
 
-                Dim codelensService = New RemoteCodeLensReferencesService(workspace.GlobalOptions)
+                Dim codelensService = workspace.Services.GetRequiredService(Of ICodeLensReferencesService)()
 
                 Dim originalDocument = workspace.Documents.First(Function(d) d.AnnotatedSpans.ContainsKey("Original"))
 
@@ -145,8 +146,9 @@ class {|Definition:C1|}
 
                 Dim root = Await startDocument.GetSyntaxRootAsync()
                 Dim node = root.FindNode(originalDocument.AnnotatedSpans("Original").First()).AncestorsAndSelf().OfType(Of ClassDeclarationSyntax).First()
-                Dim results = Await codelensService.FindReferenceLocationsAsync(workspace.CurrentSolution, startDocument.Id, node, CancellationToken.None)
-                Assert.True(results.HasValue)
+                Dim unmappedResults = Await codelensService.FindReferenceLocationsAsync(workspace.CurrentSolution, startDocument.Id, node, CancellationToken.None)
+                Assert.True(unmappedResults.HasValue)
+                Dim results = Await codelensService.MapReferenceLocationsAsync(workspace.CurrentSolution, unmappedResults.Value, ClassificationOptions.Default, CancellationToken.None)
 
                 Dim definitionDocument = workspace.Documents.First(Function(d) d.AnnotatedSpans.ContainsKey("Definition"))
                 Dim definitionText = Await workspace.CurrentSolution.GetDocument(definitionDocument.Id).GetTextAsync()
@@ -158,7 +160,7 @@ class {|Definition:C1|}
 
                 Dim actual = New List(Of (String, LinePosition, String))
 
-                For Each result In results.Value
+                For Each result In results
                     actual.Add((result.FilePath, New LinePosition(result.LineNumber, result.ColumnNumber), result.ReferenceLineText))
                 Next
